@@ -3,13 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@types/jest": "^20.0.7",
-    "@types/node": "^8.0.22",
-    "@types/react": "^16.0.4",
-    "@types/react-dom": "^16.0.4",
-    "react": "^16.0.4",
-    "react-dom": "^16.0.4",
-    "react-scripts-ts": "2.14.0"
+    "react": "^16.5.0",
+    "react-dom": "^16.5.0",
+    "react-scripts-ts": "2.17.0"
   },
   "scripts": {
     "start": "react-scripts-ts start",
@@ -24,8 +20,12 @@
     }
   },
   "devDependencies": {
-    "prettier": "^1.5.3",
-    "tslint-config-prettier": "^1.3.0",
-    "typescript": "^2.8.1"
+    "@types/jest": "^23.3.2",
+    "@types/node": "^10.9.4",
+    "@types/react": "^16.4.13",
+    "@types/react-dom": "^16.0.7",
+    "prettier": "^1.14.2",
+    "tslint-config-prettier": "^1.15.0",
+    "typescript": "^3.0.3"
   }
 }
