@@ -1,19 +1,19 @@
-import * as React from "react";
-import "./App.css";
+import * as React from 'react';
+import './App.css';
 
-const logo = require("./logo.svg");
+import logo from './logo.svg';
 
-class App extends React.Component<{}, { api_message: string }> {
+class App extends React.Component<{}, { apiMessage: string }> {
   constructor(props: object) {
     super(props);
 
-    this.state = { api_message: "" };
+    this.state = { apiMessage: "" };
   }
 
   public componentDidMount() {
-    fetch("/api").then(r => r.text()).then(api_message => {
+    fetch("/api").then(r => r.text()).then(apiMessage => {
       this.setState({
-        api_message
+        apiMessage
       });
     });
   }
@@ -21,15 +21,15 @@ class App extends React.Component<{}, { api_message: string }> {
   public render() {
     return (
       <div className="App">
-        <div className="App-header">
+        <header className="App-header">
           <img src={logo} className="App-logo" alt="logo" />
-          <h2>Welcome to React</h2>
-        </div>
+          <h1 className="App-title">Welcome to React</h1>
+        </header>
         <p className="App-intro">
           To get started, edit <code>src/App.tsx</code> and save to reload.
         </p>
         <p>
-          {this.state.api_message}
+          {this.state.apiMessage}
         </p>
       </div>
     );
